# recruitment-system/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Recruitment System API",
        "description": "API documentation for the Recruitment System",
        "contact": {
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get a list of active jobs",
                "operationId": "9538d30d530d4c3680ea31fe228944e2",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Job"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Create a new job",
                "operationId": "36deb2a6480382cf8385911c10aeebf0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "requirements",
                                    "company_name"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Software Engineer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Develop and maintain web applications."
                                    },
                                    "requirements": {
                                        "type": "string",
                                        "example": "PHP, Laravel, Vue.js"
                                    },
                                    "company_name": {
                                        "type": "string",
                                        "example": "Tech Corp"
                                    },
                                    "min_salary": {
                                        "type": "integer",
                                        "example": 80000
                                    },
                                    "max_salary": {
                                        "type": "integer",
                                        "example": 120000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Job created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "job": {
                                            "$ref": "#/components/schemas/Job"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/jobs/{job}": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get a specific job",
                "operationId": "9c31b377241c635bf2819779f358d414",
                "parameters": [
                    {
                        "name": "job",
                        "in": "path",
                        "description": "ID of the job",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Job"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Update a job",
                "operationId": "e70ad6e7b080085fa9d6faebb58ac0b2",
                "parameters": [
                    {
                        "name": "job",
                        "in": "path",
                        "description": "ID of the job",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "requirements",
                                    "company_name",
                                    "status"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "requirements": {
                                        "type": "string"
                                    },
                                    "company_name": {
                                        "type": "string"
                                    },
                                    "min_salary": {
                                        "type": "integer"
                                    },
                                    "max_salary": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "closed"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Job updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "job": {
                                            "$ref": "#/components/schemas/Job"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Delete a job",
                "operationId": "1c0f94afff104129e896097b961e009c",
                "parameters": [
                    {
                        "name": "job",
                        "in": "path",
                        "description": "ID of the job",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Job deleted successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/my-posted-jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get jobs posted by the authenticated user",
                "operationId": "b96e7eeb13e6dc2cc93315935bf07bde",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Job"
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Job": {
                "title": "Job",
                "description": "Job model",
                "properties": {
                    "id": {
                        "description": "Job ID",
                        "type": "integer",
                        "format": "int64"
                    },
                    "title": {
                        "description": "Job title",
                        "type": "string"
                    },
                    "description": {
                        "description": "Job description",
                        "type": "string"
                    },
                    "requirements": {
                        "description": "Job requirements",
                        "type": "string"
                    },
                    "company_name": {
                        "description": "Company name",
                        "type": "string"
                    },
                    "min_salary": {
                        "description": "Minimum salary",
                        "type": "integer"
                    },
                    "max_salary": {
                        "description": "Maximum salary",
                        "type": "integer"
                    },
                    "status": {
                        "description": "Job status",
                        "type": "string",
                        "enum": [
                            "active",
                            "closed"
                        ]
                    },
                    "user_id": {
                        "description": "ID of the user who posted the job",
                        "type": "integer"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}